requirements.txt
appdirs==1.4.3
Babel==2.4.0
dj-database-url==0.4.2
Django==1.10.6
django-bootstrap3==8.2.1
django-cleanup==0.4.2
django-model-utils==2.6.1
django-phonenumber-field==1.2.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
olefile==0.44
packaging==16.8
phonenumberslite==8.3.3
Pillow==4.0.0
psycopg2==2.7
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
